refactor: wrapper 삭제
