+++ apps/ai_api/app/api/error_normalizers/__init__.py
@@ -1,0 +1,1 @@
+"""라우트에서 발생한 예외를 AppError로 정규화하는 모듈 모음."""
