+++ apps/ai_api/app/api/errors.py
@@ -2,9 +2,6 @@ import logging
 from enum import StrEnum
 from typing import Any
 
-from fastapi import HTTPException
-
-
 logger = logging.getLogger(__name__)
 
 
@@ -44,92 +41,3 @@ class AppError(Exception):
         self.cause = cause
         super().__init__(detail)
 
-
-def _error_payload(
-    *,
-    code: ErrorCode,
-    message: str,
-    status: int,
-    layer: str,
-    phase: str,
-    retryable: bool,
-    context: dict[str, Any] | None = None,
-    cause: Exception | None = None,
-) -> dict[str, Any]:
-    payload: dict[str, Any] = {
-        "error": {
-            "code": code,
-            "message": message,
-            "status": status,
-            "layer": layer,
-            "phase": phase,
-            "retryable": retryable,
-            "context": context or {},
-        }
-    }
-    if cause is not None:
-        payload["error"]["cause"] = {
-            "type": type(cause).__name__,
-            "detail": str(cause),
-        }
-    return payload
-
-
-def to_http_exception(exc: Exception) -> HTTPException:
-    if isinstance(exc, HTTPException):
-        if isinstance(exc.detail, dict) and "error" in exc.detail:
-            return exc
-        return exc
-    if isinstance(exc, AppError):
-        return HTTPException(
-            status_code=exc.status_code,
-            detail=_error_payload(
-                code=exc.code,
-                message=exc.detail,
-                status=exc.status_code,
-                layer=exc.layer,
-                phase=exc.phase,
-                retryable=exc.retryable,
-                context=exc.context,
-                cause=exc.cause,
-            ),
-        )
-    if isinstance(exc, FileNotFoundError):
-        return HTTPException(
-            status_code=404,
-            detail=_error_payload(
-                code=ErrorCode.RESOURCE_NOT_FOUND,
-                message=str(exc),
-                status=404,
-                layer="infrastructure",
-                phase="resource_load",
-                retryable=False,
-                cause=exc,
-            ),
-        )
-    if isinstance(exc, ValueError):
-        return HTTPException(
-            status_code=400,
-            detail=_error_payload(
-                code=ErrorCode.INVALID_REQUEST,
-                message=str(exc),
-                status=400,
-                layer="usecase",
-                phase="validation",
-                retryable=False,
-                cause=exc,
-            ),
-        )
-    logger.exception("Unhandled exception during API request")
-    return HTTPException(
-        status_code=500,
-        detail=_error_payload(
-            code=ErrorCode.INTERNAL_SERVER_ERROR,
-            message="Internal Server Error",
-            status=500,
-            layer="api",
-            phase="unhandled",
-            retryable=False,
-            cause=exc,
-        ),
-    )
