+++ apps/ai_api/app/presentation/requests/__init__.py
@@ -1,11 +1,7 @@
-from app.presentation.requests.inference_request_dto import (
-    InferenceUploadRequestDto,
-    parse_inference_upload_request,
-)
+from app.presentation.requests.inference_request_dto import InferenceUploadRequestDto
 from app.presentation.requests.sionna_request_dto import SionnaRunRequestDto
 
 __all__ = [
     "SionnaRunRequestDto",
     "InferenceUploadRequestDto",
-    "parse_inference_upload_request",
 ]
