+++ apps/ai_api/app/presentation/responses/__init__.py
@@ -1,7 +1,3 @@
-from app.presentation.responses.inference_response_dto import (
-    build_unet_response,
-    build_yolo_response,
-)
 from app.presentation.responses.sionna_response_dto import (
     SionnaRunResponseDto,
     to_sionna_response,
@@ -10,6 +6,4 @@ from app.presentation.responses.sionna_response_dto import (
 __all__ = [
     "SionnaRunResponseDto",
     "to_sionna_response",
-    "build_unet_response",
-    "build_yolo_response",
 ]
