+++ apps/ai_api/app/usecases/run_unet_usecase.py
@@ -1,10 +1,19 @@
 from app.presentation.requests.inference_request_dto import InferenceUploadRequestDto
-from app.presentation.responses.inference_response_dto import build_unet_response
+from packages.contracts.inference import UnetInferenceResponse
 
 
-def run_unet_usecase(req: InferenceUploadRequestDto, runner):
+def run_unet_usecase(req: InferenceUploadRequestDto, runner) -> UnetInferenceResponse:
     prob_map_path, metrics = runner(req.file_id, req.image_bytes, req.filename)
-    return build_unet_response(req.file_id, prob_map_path, metrics)
+    return UnetInferenceResponse(
+        status="ok",
+        task="unet",
+        fileId=req.file_id,
+        output={
+            "wallProbNpyPath": prob_map_path,
+            "wallProbOverlayPath": metrics["overlayPath"],
+        },
+        metrics=metrics,
+    )
 
 
 __all__ = ["run_unet_usecase"]
