+++ apps/ai_api/app/usecases/run_yolo_usecase.py
@@ -1,10 +1,16 @@
 from app.presentation.requests.inference_request_dto import InferenceUploadRequestDto
-from app.presentation.responses.inference_response_dto import build_yolo_response
+from packages.contracts.inference import YoloInferenceResponse
 
 
-def run_yolo_usecase(req: InferenceUploadRequestDto, runner):
+def run_yolo_usecase(req: InferenceUploadRequestDto, runner) -> YoloInferenceResponse:
     detections, preview_path, metrics = runner(req.file_id, req.image_bytes, req.filename)
-    return build_yolo_response(req.file_id, detections, preview_path, metrics)
+    return YoloInferenceResponse(
+        status="ok",
+        task="yolo",
+        fileId=req.file_id,
+        output={"detections": detections, "previewPath": preview_path},
+        metrics=metrics,
+    )
 
 
 __all__ = ["run_yolo_usecase"]
